大改 SDK，使用 Cocoapods.

+++ UpYunSDK2.0Demo/UpYunSDK2.0Demo/ViewController.m
@@ -53,7 +53,7 @@
         [alert show];
         NSLog(@"%@",error);
     };
-    uy.progressBlocker = ^(CGFloat percent,long long requestDidSendBytes)
+    uy.progressBlocker = ^(CGFloat percent, NSUInteger requestDidSendBytes)
     {
         [_pv setProgress:percent];
     };
